Add `internal` reference
References are needed for other parts of our tooling, such as generating links. However in some cases we don't want to show these references to the user, often because they are already implicitly included

This proposals adding a new `.internal` option on `PromptReference`. When set to true, our UI will not show the reference to the user but the reference will still be accessible to other tooling