feat: add a `TokenLimit` element
This adds a `TokenLimit` element that can be used to set a hard cap on
the number of tokens that can be consumed by a prompt or part of a
prompt. Using it is fairly straightforward:

```tsx
class PromptWithLimit extends PromptElement {
	render() {
		return (
			<UserMessage>
				<TokenLimit max={1000}>{/* Your elements here! */}</TokenLimit>
			</UserMessage>
		);
	}
}
```

`TokenLimit` subtrees are pruned before the prompt gets pruned. As you
would expect, the `PromptSizing` of child elements inside of a limit
reflect the reduced budget. If the `TokenLimit` would get `tokenBudget`
smaller than its maximum via the usual distribution rules, then that's
given it child elements instead (but pruning to the `max`
value still happens.)