refactor: add a 'raw' output type, and tie types to tokenizers
Previously this library output OpenAI types as its "raw" type, but this
is not desirable if we ever support models with features that aren't a
subset of OpenAI's features. We also had tokenizer struggles when adding
images due to restrictive types and muddling of VS Code and OAI types.

This commit makes the basic output of the library to be a "raw" chat
message type, that can be converted into VS Code or OpenAI messages,
potentially with loss. The ITokenizer declares its mode, and that mode
is used to convert messages for counting and as `renderPrompt` output.

This is a breaking changes especially if consumers use the
PromptRenderer directly, but it should be easy to adopt with the helper
toVSCode / toOpenAI methods.